issue-88-upgrade - Revising code and documentation some more.

+++ java/com/almworks/sqlite4java/SQLiteConstants.java
@@ -160,11 +160,10 @@ public interface SQLiteConstants {
   int SQLITE_OPEN_PRIVATECACHE = 0x00040000;
   int SQLITE_OPEN_WAL = 0x00080000;
 
-  //Flags intended for "prepFlags" parameter of the sqlite3_prepare_v3(). More may exist in future sqlite versions.
   int SQLITE_PREPARE_PERSISTENT = 0x01;
 
-  long SQLITE_WIN32_DATA_DIRECTORY_TYPE = 1;
-  long SQLITE_WIN32_TEMP_DIRECTORY_TYPE = 2;
+  int SQLITE_WIN32_DATA_DIRECTORY_TYPE = 1;
+  int SQLITE_WIN32_TEMP_DIRECTORY_TYPE = 2;
 
   // SQLITE_LIMIT_* identify class of constructs to be size limited
   int SQLITE_LIMIT_LENGTH = 0;
